third_party/blink/web_tests/external/wpt/shadow-dom/focus/click-focus-delegatesFocus-tabindex-varies-expected.txt

This is a testharness.js-based test.
FAIL click on host with delegatesFocus, #aboveSlot tabindex = 2, #slot and #slotted tabindex = 1 assert_equals: expected Element node <div tabindex="2">aboveSlot</div> but got null
Harness: the test ran to completion.
